Refactor slidepad

+++ test/test_dac.h
@@ -12,17 +12,17 @@ typedef struct TestDAC
 {
     wby_idac_t parent;
     TestDACState state;
-    uint16_t value;
+    uint8_t value;
 } TestDAC;
 
-static void test_dac_sink(wby_idac_t *parent, uint16_t value)
+static void test_dac_sink(wby_idac_t *parent, uint8_t value)
 {
     TestDAC *self = (TestDAC *)parent;
     self->state = TEST_DAC_SINK;
     self->value = value;
 }
 
-static void test_dac_source(wby_idac_t *parent, uint16_t value)
+static void test_dac_source(wby_idac_t *parent, uint8_t value)
 {
     TestDAC *self = (TestDAC *)parent;
     self->state = TEST_DAC_SOURCE;
